+++ CHANGES.md
@@ -5,7 +5,8 @@
 
 ## 0.1.11.dev0 (Work In Progress)
 
-- 
+- add `ipyexperiments.utils.ipython` module with workaround for ipython memory leak on exception problem.
+- add `ipyexperiments.utils.mem` module with memory debug helper utils.
 
 
 ## 0.1.10 (2019-01-18)